requirements.txt
langchain
langchain_community
pandasai
ollama 
streamlit
groq
langchain-groq
python-dotenv
